tools: gateway: extract `LocalRpcServer`
Extract this helper class so it can be re-used by other tools.
